Work on the services and controllers and searching with pagination
1. Find workers, clients and assignments by pagination
2. Created all basic service and controller
3. Renaming fields in classes
4. Added dependency for test in pom.xml

+++ src/main/java/com/example/cleaningcompanyplanner/worker/WorkerRepository.java
@@ -2,5 +2,5 @@ package com.example.cleaningcompanyplanner.worker;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkerRepository extends JpaRepository<Worker,Long> {
+public interface WorkerRepository extends JpaRepository<Worker, Integer> {
 }
